+++ src/apis/plugins_api.rs
@@ -202,7 +202,7 @@ pub async fn enable_plugin(configuration: &configuration::Configuration, plugin_
 }
 
 /// Gets all plugins from the marketplace server, merging data from locally installed plugins as well as prepackaged plugins shipped with the server.  ##### Permissions Must have `manage_system` permission.  __Minimum server version__: 5.16 
-pub async fn get_marketplace_plugins(configuration: &configuration::Configuration, page: Option<i32>, per_page: Option<i32>, filter: Option<&str>, server_version: Option<&str>, local_only: Option<bool>) -> Result<Vec<crate::models::MarketplacePlugin>, Error<GetMarketplacePluginsError>> {
+pub async fn get_marketplace_plugins(configuration: &configuration::Configuration, page: Option<i64>, per_page: Option<i64>, filter: Option<&str>, server_version: Option<&str>, local_only: Option<bool>) -> Result<Vec<crate::models::MarketplacePlugin>, Error<GetMarketplacePluginsError>> {
     let local_var_configuration = configuration;
 
     let local_var_client = &local_var_configuration.client;
