+++ src/models/data_retention_policy_for_channel.rs
@@ -18,7 +18,7 @@ pub struct DataRetentionPolicyForChannel {
     pub channel_id: Option<String>,
     /// The number of days a message will be retained before being deleted by this policy.
     #[serde(rename = "post_duration", skip_serializing_if = "Option::is_none")]
-    pub post_duration: Option<i32>,
+    pub post_duration: Option<i64>,
 }
 
 impl DataRetentionPolicyForChannel {
